Changing files and connection information
update on schemas and models
