+++ server/models copy/Comment.js
@@ -1,0 +1,32 @@
+const { Schema, Types } = require("mongoose");
+
+const commentSchema = new Schema(
+  {
+    commentId: {
+      type: Schema.Types.ObjectId,
+      default: () => new Types.ObjectId(),
+    },
+    commentBody: {
+      type: String,
+      required: true,
+      maxlength: 280,
+    },
+    username: {
+      type: String,
+      required: true,
+    },
+    createdAt: {
+      type: Date,
+      get: (date) => date.toLocaleDateString(),
+      default: Date.now,
+    },
+  },
+  {
+    toJSON: {
+      getters: true,
+    },
+    id: false,
+  }
+);
+
+module.exports = commentSchema;
