+++ server/models copy/User.js
@@ -23,12 +23,13 @@ const userSchema = new Schema({
     required: true,
     minlength: 5,
   },
-  // bugs: [
-  //   {
-  //     type: Schema.Types.ObjectId,
-  //     ref: "Bug",
-  //   },
-  // ],
+  bugs: [
+    {
+      type: Schema.Types.ObjectId,
+      ref: "Bug",
+    },
+  ],
+  image: String,
 });
 
 userSchema.pre('save', async function (next) {
