+++ server/models copy/index.js
@@ -1,0 +1,10 @@
+const User = require("./User");
+const Bug = require("./Bug");
+const Game = require ("./Game")
+//comment is a schema, that is why it is not included in index.js
+module.exports = {
+  User,
+  Comment,
+  Game,
+  Bug,
+};
