--- server/models/index.js
@@ -1,5 +1,0 @@
-const User = require("./User");
-
-module.exports = {
-  User,
-};
